backend/requirements.txt.txt
fastapi
uvicorn
python-dotenv
pypdf
langchain
langchain-openai
